Move KotlinHubTextField to a separate folder
